backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
supabase==2.3.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
httpx==0.26.0
pdfplumber==0.10.3
pypdf==4.0.1
openai==1.12.0
pytest==7.4.4
pytest-asyncio==0.23.3
ruff==0.1.11
